Hurray for feat wanting to do things differently.
Because it just ends up doing it wrong.

In this case, someone thought that depth was a better name than where,
and that the arguments should be positive instead of negative.

But it doesn't properly map one to the other, and hence everything
breaks.  So it wasn't only different for no good reason except than
being different, it was also wrong.  Another hour wasted.