Customize futher the Resolver used by feat.
There is a problem with using twisted.names.client in twisted 10.2. If the
dns queries return both AAAA and A type of entries, the AAAA type (IPv6) is
favored. In the end the result of reactor.resolve() is a IPv6 address.
Unfortunately this doesn't work well with twisted.internet.tcp.Connector,
which expects IPv4 address to connect to.

This fix is motivated by the problems with connecting to maps.googleapis.com.