Refactor the bootstrap.
When daemonizing, the original process waits for the daemon process to
finish the agency intialization before it returns. It exists with zero
status only if agency has intialized successfuly. The logging of this
process is done to feat-service.log. In case of failure during
initialization, the message is transmitted to the master process.

Also always log the file name we switch logging to to feat-service.log.