Increase timeout for starting agent to 20 seconds.
The motivation here is the behaviour of frf. Comparing the logs of the
master and the slave agencies I can see that it takes 7 seconds there
between master calling "feat -X ..." and the slave logging the first line
to the log. There is simply not enough time for the slave to start up.