Add refcount to PersonalBinding() object.
The motivation is as follows. It was possible that two interests share the same binding
for example when the agent is in 'lobby' when creates the interest and calls
bind_to_lobby() afterwards. Than, if one binding was removed (because of shard change)
the banding to RabittMQ was removed as well.